Add code to modules file
